Move the AllSites component to blocks
