+++ assets/stylesheets/_components.scss
@@ -20,6 +20,7 @@
 @import 'account-recovery/reset-code-validation/style';
 @import 'account-recovery/style';
 @import 'auth/style';
+@import 'blocks/all-sites/style';
 @import 'blocks/app-banner/style';
 @import 'blocks/app-promo/style';
 @import 'blocks/author-compact-profile/style';
@@ -334,7 +335,6 @@
 @import 'me/social-login/style';
 @import 'me/two-step/style';
 @import 'my-sites/ads/style';
-@import 'my-sites/all-sites/style';
 @import 'my-sites/all-sites-icon/style';
 @import 'my-sites/checklist/style';
 @import 'my-sites/current-site/style';
